backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-settings==2.2.1
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.3.2
python-dotenv==1.0.1
